refactor(hub): move bundle into a new with-share dimension (8th)
Splits 'copies that leave the timeline' by who ends up holding them: with-fork
(archive·shadow·snapshots — in-custody views/checkpoints) vs with-share (bundle
+ extract/adopt — the data leaves, re-keyed, adoptable) vs with-party (access
shared with people). bundle's defining trait is data-portability/transfer, not
forking a copy.

Pure move: git mv with-fork/bundle → with-share/bundle; rewrite the importers
(same-depth swap) + the 1 sibling recompute (snapshots/engine.ts) + the tsup
entry source path. Public subpath @noy-db/hub/bundle byte-identical (29 exports
unchanged). typecheck ✓ lint ✓ test ✓ (2942) arch ✓ build ✓. north-star doc
updated to 8 dimensions.
