+++ packages/hub/__tests__/custody-api.test.ts
@@ -28,7 +28,7 @@
 import { describe, it, expect, beforeEach } from 'vitest'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/types.js'
 import { ConflictError, PermissionDeniedError, ReadOnlyError, PartitionExtractionError } from '../src/errors.js'
-import { extractPartition } from '../src/with-fork/bundle/extract-partition.js'
+import { extractPartition } from '../src/with-share/bundle/extract-partition.js'
 import { createNoydb } from '../src/noydb.js'
 import type { Noydb } from '../src/noydb.js'
 import { withHistory } from '../src/with-commit/history/index.js'
