+++ packages/hub/__tests__/decrypt-partition.test.ts
@@ -15,8 +15,8 @@ import { createNoydb } from '../src/noydb.js'
 import type { Noydb } from '../src/noydb.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/types.js'
 import { ConflictError } from '../src/errors.js'
-import { extractPartition } from '../src/with-fork/bundle/extract-partition.js'
-import { decryptExtractedPartition } from '../src/with-fork/bundle/decrypt-partition.js'
+import { extractPartition } from '../src/with-share/bundle/extract-partition.js'
+import { decryptExtractedPartition } from '../src/with-share/bundle/decrypt-partition.js'
 
 function memory(): NoydbStore {
   const store = new Map<string, Map<string, Map<string, EncryptedEnvelope>>>()
