+++ packages/hub/src/collection.ts
@@ -67,7 +67,7 @@ import { buildDescription, deriveZodFields, type CollectionDescription, type Des
 import { buildJsonSchema } from './with-shape/introspection/json-schema.js'
 import type { CollectionConfig } from './with-shape/introspection/types.js'
 import { Lru, parseBytes, estimateRecordBytes, type LruStats } from './cache/index.js'
-import { generateULID } from './with-fork/bundle/ulid.js'
+import { generateULID } from './with-share/bundle/ulid.js'
 import type { PresenceHandle, PresenceHandleOpts } from './with-party/team/presence.js'
 import { NO_SYNC, type SyncStrategy } from './with-party/team/sync-strategy.js'
 import type { BlobSet } from './with-shape/blobs/blob-set.js'
