+++ packages/hub/src/with-audit/consent/consent.ts
@@ -50,7 +50,7 @@
  */
 import type { EncryptedEnvelope, NoydbStore } from '../../types.js'
 import { encrypt, decrypt } from '../../crypto.js'
-import { generateULID } from '../../with-fork/bundle/ulid.js'
+import { generateULID } from '../../with-share/bundle/ulid.js'
 
 /** Reserved collection for consent-audit entries. */
 export const CONSENT_AUDIT_COLLECTION = '_consent_audit'
