+++ packages/hub/src/with-share/bundle/adopt-partition.ts
