+++ packages/hub/src/with-share/bundle/extract-partition.ts
