+++ packages/hub/src/with-share/bundle/walk-closure.ts
